requirements.txt
opencv-python==4.11.0.86
pillow==11.2.1
requests==2.32.3